+++ libft/ft_strlen.c
@@ -6,15 +6,15 @@
 /*   By: chyeok <chyeok@student.42seoul.kr>         +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/03/14 13:27:29 by chyeok            #+#    #+#             */
-/*   Updated: 2022/03/15 19:08:48 by chyeok           ###   ########.fr       */
+/*   Updated: 2022/05/20 04:27:29 by chyeok           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-#include "libft.h"
+#include <stddef.h>
 
-int	ft_strlen(const char *str)
+size_t	ft_strlen(const char *str)
 {
-	int	count;
+	size_t	count;
 
 	count = 0;
 	while (str[count])
